fix: avoid wrapping around seconds from epoch
Now `Token::sign` panics if the seconds from UNIX epoch cannot be
represented in a i64. Previously this would have wrapped around,
possibly creating a _very old_ JWT.

AFAIK, this is _technically_ a breaking change, but in reality this
should only affect systems maliciously set with a very future
SystemTime. Therefore I think that it should not require a major version
bump.